requirements.txt
streamlit>=1.38.0
openai>=1.51.2
pandas>=2.0.0
sqlalchemy>=2.0.0
sqlmodel>=0.0.14
plotly>=5.17.0
altair>=5.1.0
wordcloud>=1.9.2
numpy>=1.24.0
python-dotenv>=1.0.1
nltk>=3.8.1
Pillow>=10.0.0
requests>=2.32.3
yake>=0.4.8
